[#99] Rename: rename mocks/user to mocks/users folder
