+++ src/mocks/handlers.ts
@@ -1,5 +1,5 @@
 import planHandler from '@/mocks/plans/handler';
-import userHandler from '@/mocks/user/handler';
+import userHandler from '@/mocks/users/handler';
 import workoutHandler from '@/mocks/workouts/handler';
 
 const handlers = [...userHandler, ...planHandler, ...workoutHandler];
